- Fixed: script cacher ignored newlines, causing wrong script line context data in server logs. - Fixed: custom triggers not firing. - Fixed: CUOClientVersion not parsing correctly new string format. - Fixed: TOOLTIP Dialog keyword sending the cliloc number in hex notation if ini DecimalVariables == 0 (unsupported by client, making some of them crash). - Fixed: some integer data type size inconsistencies. - Fixed: initialization order fiasco for CAccounts at server shutdown. - Changed: Str_To* functions to return a std::optional value, in order to get the best error handling (before, it depended on errno and there were ambiguous return values for invalid or "0" zero strings. - Changed: Refactored CUOMobTypes. - Changed: replaced localtime function with thread-safe alternatives, avoiding data race conditions. - Added: stypecast.h. It has utilities to cast numbers with strong guarantees and error checking at compile time, to avoid undefined behaviors, involuntary wrong sign/size cast or overflowing.

+++ src/network/CIPHistoryManager.h
@@ -25,7 +25,7 @@ struct HistoryIP
     int m_connected;
     bool m_blocked;
     int m_ttl;
-    size_t m_connectionAttempts; // since i remember of this IP
+    int64 m_connectionAttempts; // since i remember of this IP
     int64 m_timeLastConnectedMs;
     int64 m_blockExpire;
     int m_pingDecay;
